디자인 컴포넌트 분리
